background_removal/requirements.txt
numpy==1.19.5
Pillow==9.0.1
scikit_image==0.17.2
skimage==0.0
torch==1.10.2
torchvision==0.11.3
